Refactor views.
